feat: update get booking API with change price fields in booking detail to float
- Update RawPrice and Price fields in GetBookingDetailItem struct from int to float64 for better precision in monetary values.
- Adjust service logic to handle float values for booking details accordingly.

+++ internal/model/admin/booking/get.go
@@ -35,8 +35,8 @@ type GetTimeSlot struct {
 type GetBookingDetailItem struct {
 	ID       string     `json:"id"`
 	Service  GetService `json:"service"`
-	RawPrice int        `json:"rawPrice"`
-	Price    int        `json:"price"`
+	RawPrice float64    `json:"rawPrice"`
+	Price    float64    `json:"price"`
 }
 
 type GetService struct {
